Refactor payment-related logic and enhance UX for options.
Refactored `PaymentHelper` to centralize basket and savings logic shared across integrations. Updated controllers like `LeaseController` and `FinanceController` to use this helper. Renamed `ChoosePaymentOption` to `Options` for clarity and improved frontend UX with dynamic offer handling.

+++ database/migrations/0000_00_00_000000_create_payment_offers_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name')->nullable();
             $table->string('type');
             $table->integer('payment_survey_id')->index();
-            $table->integer('amount');
+            $table->decimal('amount', 8, 1);
             $table->integer('payment_provider_id')->index();
             $table->decimal('apr', 8, 1)->nullable();
             $table->integer('term');
